Add new extra "all", recommended install path
As discused in #222, the potential loss of feature that disabling
matplotlib by default is not nice.

To avoid that, add a new install extra meta-feature called "all", which
will aggregate all extras one might want (currently just matplotlib).
This becomes the new recommended way to install, so no loss of feature
for those who use it, but allow the lighter option by disabling the all.

Update docs accordingly (may be updating it a little strong).
New docs entry for matplotlib/numpy support to align with plantuml.