needs: Remove some of the extra IDs in the output
Inside the LaTeX generator, it looks for the attributes ids and refid
for creating link targets.  When a new layer with a <container> wrapper
was added in commit e51fc1f, for an unknown reason this triggered these
additional IDs to generate links within the LaTeX output.

A test case has been added to ensure only one hyperref within LaTeX is
encountered as well.

(https://github.com/useblocks/sphinx-needs/issues/808)