latex output: Fix LaTeX output format for needs tables
This adds the `colwidths-given` class to the table generated with a
simple needs table.  The LaTeX output then actually checks the
<colspec> blocks for each column and respects the widths specified.

Additionally, the need table was originally generating a <line-block>
context and populating it with <line> elements.  This ends up adding
extra spacing into the output for whatever reason and took up a lot of
vertical space for each need in the rendered LaTeX PDF output.  This has
been removed and instead each line is placed into a new paragraph
context via nodes.paragraph().  This renders the output correctly, and
the HTML side is captured via CSS to ensure that there are no margins
around the paragraph.

Other attempts to solve the rendering didn't work.  If I moved the class
into a nodes.inline(...) element, the LaTeX renderer doesn't like having
a nested paragraph inside of a class defined via `inline(...)`.  Placing
a pair of newlines, defining a new paragraph that way, also yielded the
same error from the LaTeX renderer.