🐛 Fix adding sections to hidden needs
Previously hidden needs would not be added to the doctree.
However, this meant that the `add_sections` transform did not add section related data fields to the need data.
(this caused errors in the docs builds when filters required these fields)

The needs node is not created in the doctree, but marked with a `hidden` attribute, and these nodes are skipped in the HTML/LaTeX rendering.