🔧 Move data from global env to individual nodes
It is unnecessary to store this data centrally, which increases complexity (having to keep it in-sync) and the size of the `BuildEnvironment`.

The only node I have not done this for is `needuml`, since there is some additional interplay between the `arch` need field, which I do not fully understand yet.