🔧 Remove `content_node` field from need items (#1241)
`content_node` is different to other need fields,
in that it is a non-jsonable type, and is not expected to be used by
users in need filters.
Its only use is for `needextract` directives.

Therefore, we move it to its own separate "storage solution". For now,
they are still stored on the `BuildEnvironment`, but ultimately we may
want to pickle them to disk (similar to doctrees), since they can add a
significant memory size to the `BuildEnvironment` that can have effects
on performance.