🐛 Handle malformed `filter-func` option value
Currently malformed filter function strings would simply be ignored,
and thus the results would not be filtered,
which is especially bad if you have a large amount of needs.

This PR provides a warning for all malformed `filter-func` values,
and also returns the result as no needs,
so we do not wasted time processing them.