🔧 Improve `process_filters` function
Currently based on-top of #1255

This PR changes `process_filters`, so that we pass it the `NeedsView` rather than prematurely converting this to a list of needs.
In the future, this will allow us to improve filter performance.

We also parse it the "origin" node, rather than prematurely converting this to a location string.