♻️ Replace need dicts/lists with views (with fast filtering) (#1281)
This commit addresses issues in performance scalability of needs filtering:

- In a known user project, with ~40,000 needs, there are 5316 individual filter calls, across directives such as `needarch`, `needtable`, `needpie`, etc. Each takes on average 0.33 seconds, cumulative totalling **1751 seconds**
- Applying this PR reduces the average time to 0.033 seconds, cumulatively totalling **181 seconds**

----

The main issue with filtering (a.k.a. querying) needs, is that it requires looping over every need and executing a Python evaluation of the filter string (e.g. `id == "xxx"`), which scales with `O(N)` time.

To reduce this to `O(1)` time, for common patterns, we look to do two things:

1. More tightly control access to needs data across its lifecycle (see #1264): 
   By making needs data strictly immutable/read-only during the write/analysis phase (after it has been fully collected and post-processed), and moving access behind an abstract interface (as opposed to a standard dictionary),
   we can perform indexing on standard need fields, making value lookups `O(1)`

2. Analysing the filter string for common patterns (by parsing/analysing the Python AST), we can then utilise *index lookups* rather than *row scans* in these cases (akin to [SQL query plans](https://sqlite.org/eqp.html)).

---

One complication in creating the indexes and abstract interfaces, is that within the code base there are essentially two ways of accessing needs data:

- As a mapping of need ID to need data (now `NeedsView`)
- As a list of both needs and "expanded" parts (now `NeedsAndPartsListView`)

Particularly when it comes to e.g. `id == "xxx"`, this then has different meanings, as the former is only filtering for need IDs and the latter is filtering for both Need IDs and part IDs

---

It is of note that this will be breaking for any projects that currently attempts to mutate needs data within the write phase.
This has been mitigated by the addition of two sphinx events, which give more *precise* control for this use case:

- `needs-before-post-processing`: callbacks `func(app, needs)` are called just before the needs are post-processed (e.g. processing dynamic functions and back links)
- `needs-before-sealing`: callbacks `func(app, needs)` just after post-processing, before the needs are changed to read-only

Additionally, `env.needs_all_needs` has been replaced with `env._needs_all_needs`, since this "raw" data-structure should not be accessed directly.

The `get_needs_view` function has been added to the `sphinx_needs.api` to mitigate this, and it should perhaps be made clearer that users should NOT be accessing any `sphinx_needs` API outside this module.