🔧 Warn on unknown need keys in external/import sources
This PR centralises the declaration of what keys to omit from external/import sources, onto the `NeedsCoreFields`.
It then also uses `NeedsCoreFields` to identify unknown keys, and emits a warning if any are found.

---

One issue that was identified during this is the distinction of `title` vs `full_title`, whereby `title` is meant to be the potentially trimmed version of `full_title`.
I think storing both fields is unnecessary and confusing; trimming of the title is a presentational issue that should be handled when necessary, not stored in the data.
Also it is of note that the need directive pre-trims the title, leading to both `title` and `full_title` being trimmed