🔧 Merge defaults into config earlier
Since https://github.com/sphinx-doc/sphinx/pull/12949 it has become more apparent that sphinx detects some configuration values as changed when they have not.
This fixes most of these by merging user defined config with built-in defaults earlier in the processing.

The only thing is does not fix is config altered by adding services. This is difficult, since services are stored on the `env`, which is not initialised until agter the config is checked.