♻️ Remove back link manipulation from `needextend` (#1386)
Back links are computed at the end of the need post-processing, after
`needextend` have been applied.

Back links should always be in-sync with forward links, therefore it
doesn't make sense to modify back links in this way.

Thankfully, the current availability of back link options in
`needextend` directives is not documented, and so are unlikely to have
been actually used.