♻️ Make `needextend` argument declarative (#1391)
The argument for `needextend` can refer to either a single need ID or
filter function.
Currently, the format cannot be known until all needs have been
processed, and it is resolved during post-processing.
This is problematic for (a) user readability, (b) improving processing
performance and issue feedback

This PR slightly modifies the argument processing to allow for two
"explicit" formats:

- `<ID>`, if the argument is enclosed in `<>` it is always processed as
a single ID
- `"filter string"`, if the argument is enclosed in `""` it is always
processed as a filter string

If the string is not enclosed, then the logic is relatively similar to
the existing:

- `ID`, if the argument is a single word (no spaces) then it is
processed as a single ID
- otherwise, it is processed as a filter string

This change should not break most uses of `needextend`, but there are
some corner-cases that would need to be adapted, e.g. a single
`is_external` would now be processed as an ID, unless it was enclosed or
expanded:

```restructuredtext
.. needextend:: "is_external"
.. needextend:: is_external is True
```