♻️ Remove `delete` from internal needs and `needs.json` (#1347)
The `:delete:` option on a need directive deletes a need before
creating/storing it, therefore it is literally impossible for it to be
anything other than `False`.
Storing the field on a need is misleading, because it suggests that the
need will be deleted (see e.g. #1383),
which is not possible with the current sphinx-needs logic.