♻️ Remove `full_title` need field and only trim generated titles
In the documentation for `needs_max_title_length`, it states:

> This option is used in conjunction with auto-generated titles as controlled by `needs_title_from_content`

However, this was not the case; it was being trimmed for all titles, irrespective of if they were auto-generated.

Trimming is now only applied to auto-generated titles (an the untrimmed title is still used for ID generation),

The existence of both `title` and `full_title` is confusing and unnecessary and so `full_title` is removed.