Allow `collapse` and `hide` in `needs_global_options`
`needs_global_options`, a.k.a need field defaults, should only be used if the user does not specifically set their value,
i.e. user specified values should take precedence over defaults.

The problem with the existing logic, was that the only check for if a value had been set was if it was "truthy".
this does not work if the user wants to explicitly set a `False` value.

For `collapse` and `hide`, this is now accounted for, and so they can be allowed as defaults.
(In future this logic maybe should be extended to all fields)