‼️ Improve needs default field application (via needs_global_options)
Previously defaults would be applied to any fields of a need with a "falsy" value, e.g. `None`, `False`, `0` , `""`, `[]`. ...

This is an issue, if the user want to specifically set fields to these values, without them being overriden by defaults.
Therefore, now defaults are only applied to fields with a missing value, denoted by `None`.

In the need directive context, this means that defaults will only be set for fields not specifically specified in the directive.

In the external/import need context, defaults will only be applied to missing fields, of fields with the value `null`.

TODO discuss restriction of predicate filtering fields, and that these field values will no longer be based on previous default applications.