‼️ Improve needs default field application (via needs_global_options) (#1478)
Previously defaults would be applied to any fields of a need with a
"falsy" value, e.g. `None`, `False`, `0` , `""`, `[]`. ...

This is an issue, if the user wants to specifically set fields to these
values, without them being overridden by defaults. Therefore, now
defaults are only applied to fields with a missing or `None` value.

In the need directive context, this means that defaults will only be set
for fields not specifically set in the directive options.

In the external/import need context, defaults will only be applied to
missing keys in the `needs.json`, or fields with the value set to
`null`.

For predicate defaults, the context passed to the expression has been
restricted to a subset of fields, and these values are no longer mutated
by previous defaults, i.e. each predicate gets the same immutable
context data.
This makes these predicates more isolated / understandable, and allows
for them potentially later being co-located with the type information.

----

‼️ Breaking

This PR may be breaking:

1. For users expecting defaults to be applied to external/import
`needs.json`, although I would suggest this was never the intention for
these defaults and that `needs.json` should, at least by default, be
seen as static ingest data which has already had defaults etc applied.

2. For users with "exotic" predicate expressions, that potentially rely
on the application of previous defaults.