♻️ Lazily assess directive options
docutils `Directive.option_spec` is designed to be a statically specified set of directive options names and validation/coercion functions.
In sphinx-needs, the `NeedDirective`, `NeedextendDirective`, and `NeedserviceDirective` directives have dynamic options, based on the schema of the need items.

Currently, the "hack" to get this to work is to dynamically set class variables on these classes. This is problematic (a) in modifying global state and (b) in having confusing logic spread across the code base.
It also means it is impossible in the future to have different need types with potentially different schemas.

This PR instead takes an approach from `sphinx/ext/autodoc/directive.py`, to allow all options to be collected, without specifying their names up-front, then analysing them within the `Directive.run` method.