♻️ Lazily assess directive options (#1482)
docutils `Directive.option_spec` is designed to be a statically
specified map of directive options names to validation/type coercion
functions.
In sphinx-needs, the `NeedDirective`, `NeedextendDirective`, and
`NeedserviceDirective` directives have dynamic options, based on the
schema of the need items.

Currently, the "hack" to get this to work is to dynamically set class
variables on these classes.
This is problematic because:

1. It modifies global state, which can be problematic for re-building in
the same Python session.
2. It leads to confusing logic, spread across the code base.
3. It also means it is impossible in the future to have different need
types with potentially different schemas.

This PR instead takes an approach from
`sphinx/ext/autodoc/directive.py`,
to allow all options to be collected, without specifying their names
up-front,
then analysing them within the `Directive.run` method.

---

‼️ Breaking

This could be breaking, for any user that relied on particular
existence/use of `Directive.option_spec` or `Directive.options`,
although this would be deemed quite an exotic use-case, and not one
intended in the API.