♻️ Store needs as `NeedItem`, rather than standard dict
Currently, need data items are stored as standard Python dicts,
which is also exposed to the user in various APIs
This is problematic as it does not allow for (a) any internal checks to ensure data is updated consistently (e.g. fields like `id` should never be changed), (b) it does not allow any data to be "hid" from the user, or presented in a different way to its internal structure.

In this PR, we instead add a specific `NeedItem` class, which closely mimics a dictionary, but also allows for additional functionality.

We also introduce a `NeedPartItem` which is distinctly different from a `NeedItem`, e.g. it should not be mutated.