♻️ Store needs as `NeedItem` / `NeedPartItem`, rather than standard `dict` (#1485)
Currently, need data items are stored as standard Python dicts, which is
also exposed to the user in various APIs and as the `needs.json`.
This is problematic as it does not allow for

1. any internal checks to ensure data is stored/updated consistently
(e.g. fields like `id` should never be changed),
2. any data to be "hid" from the user, or presented in a different way
to its internal structure.
- because the dict is a flat key/val store, it does not allow for
distinction between internal / user set extra/link fields
- we would like to improve how values that are dynamic/variant functions
are stored (prior to evaluation)
- we would also like to store source mapping data from distinct sources
(directives, external, imports) in a more coherent way.
4. deprecations of any fields in a non-breaking manner

In this PR, we instead add a specific `NeedItem` class, which closely
mimics a`dict`, but also allows for additional constraints and
functionality.

We also introduce a `NeedPartItem` which is distinctly different from a
`NeedItem`, e.g. it should not be mutated.

To ensure that no existing user code is accidentally breaking this, e.g.
by directly add dictionaries to the list of needs,
we perform a check, at the end of the post-processing stage, to make
sure all need items are instances of `NeedItem`
(and point them towards the "correct" `add_need` API)

---

‼️ Breaking

This would be breaking for any users doing "non-API" modifications /
additions to the needs data, i.e. directly adding dict items.
It should not change interactions with standard APIs like `add_need` or
filter strings, etc