♻️ Auto-compute certain need fields (#1496)
A number of need fields are actually data that is based on other data:

- `is_need` / `is_part` are based on whether the item is a `NeedItem` or
`NeedPartItem` instance
- `is_modified` is based on if `modifications > 0`
- `id_parent` / `id_complete` are based on the `id` (and part `id` if
`NeedPartItem`)
- `section_name` is the first item in `sections`
- `parent_need` is the first item in the `parent_needs` links
- `constraints` is the keys of `constraints_results`
- `constraints_passed` is derived from `constraints_results`

These are mainly there for user convenience when writing filters, but
are not necessary to specify independently, and could lead to
inconcistencies in the data.

Therefore, these fields are now auto-computed when
initialiasing/modifying the need, rather than being allowed to be
specified / modified independantly.