🧪 Improve test for need parts (#1507)
This adds more to the need part test, in order to improve testing of how
links/backlinks are recorded.
I believe it exposes a bug, whereby backlinks for the parent need are
copied to the part need, for links that are not of type "links".
This will be fixed in subsequent PR.