🐛 Avoid leaking auth credentials for ext. need warnings (#1512)
Changes for [issue
#1476](https://github.com/useblocks/sphinx-needs/issues/1476):

When a warning is triggered due to unknown keys in external sources, it
includes the URL of the source. If the URL contains basic authentication
details, those credentials will be exposed in the warning message.

FIX is simply to use the existing "clean_log" method to obfuscate the
url

Co-authored-by: cedric.dos-santos.e <cedric.dos-santos.e@thalesdigital.io>