🔧 Add `DynamicFunctionParsed` dataclass (#1514)
This will allow dynamic functions to be pre-parsed and stored, in future
changes.
The parsing is also updated, to replace deprecated classes in the `ast`
module,
and parametrized tests are added for successful/failing parsing.

Note, it would be good to just call this class `DynamicFunction`, but that is
already in use for now.