🐛 Inject build_tags to variant filter strings (#1531)
Instead of passing Sphinx tags directly into the context variable scope
for filter strings, a new variable `build_tags` is injected which is a
`set[str]`.
This avoids namespace clutter and also `NameError`s when checking for
tags.

For now, this is only available for variants. Need to discuss whether to
apply this to all filter strings.