👌 Validate all schemas during config resolution
During resolutuion of schemas coming from config `needs_extra_links`, `needs_extra_options` and `needs_schema_definitions`,
check that these schemas actually compile.
This removes the need for "bespoke" regex checking, and actually allows more regex syntaxes (but still not lookahead/lookbehind).

Additionally, do not use the `typeguard` dependency, to check schemas coming from `need_extra_options` and `need_extra_links` and instead.
The replacement checks are "simpler", but we need to remove `typeguard`, because it is not compatible with python 3.14, and I believe these simpler checks, together with the schema validation, will suffice.
Note, validation of `SchemasRootType` is still using `typeguard` for now, so we can not remove the dependency, but this can be done in a followup PR.