♻️ Remove typeguard dependency (#1597)
The `typeguard` dependency is a blocker to supporting Python 3.14
[#1566](https://github.com/useblocks/sphinx-needs/issues/1566#issuecomment-3626096968)
Currently, its only use is to validate configuration for need schemas

In this PR, we replace this functionality with validating the
configuration against JSON schemas, using the existing `jsonschema-rs`
dependency.
We generate these schemas via converting the `TypedDict` configuration
to `pydantic` models first.
Note that we only use `pydantic` as a test dependency, to auto-generate
the schemas and ensure they are up-to-date, rather than as a runtime
dependency, since we don't want to introduce an additional dependency
(with version constraints) that would make it more difficult for
sphinx-needs to integrate into people's existing Python environments.